move two properties from OsmPrimitive to Way to save some memory (we have a lot less Ways than Nodes)
add some more global on/off switches for drawing features (fillareas, strokes, icons, annotate) where you can set the maximum distance as displayed on the map. Feature (fillarea, ...) will only be drawn if current distance scaler as shown on the top left of the map is smaller than the setting. The defaults are 100000m which means "always shown".

git-svn-id: http://josm.openstreetmap.de/svn/trunk@1270 0c6e7542-c601-0410-84e7-c038aed88b3b