memory optimizations for Node & WayPoint (Patch by Gubaer, modified)
The field 'proj' in CachedLatLon is a waste of memory. For the 2 classes where this has the greatest impact, the cache for the projected coordinates is replaced by 2 simple double fields (east & north). On projection change, they have to be invalidated explicitly. This is handled by the DataSet & the GpxLayer.


git-svn-id: http://josm.openstreetmap.de/svn/trunk@4126 0c6e7542-c601-0410-84e7-c038aed88b3b