fix #6556
When the prefs are opened for the very first time, the window is now at most
as large as the parent window. After that, the dimensions are stored in JOSM's
preferences, so it won't work for existing installations.

As an additional measure, the dialog cannot be larger than the screen dimension
and as far as my testing goes this setting overwrites the stored dimension.
Unfortunately there is no way to determine the space occupied by application
launchers/taskbars across all systems and the pref dialog may be hidden behind
them.




git-svn-id: http://josm.openstreetmap.de/svn/trunk@4248 0c6e7542-c601-0410-84e7-c038aed88b3b