add session support for marker layers (see #4029)
The data is exported to a separate GPX file that contains one waypoint for each marker.
This is not very elegant, because most of the time, all the info is already contained in the original GPX File.
However, when dealing with audio markers, they can be synchronized, or additional markers are added
at certain playback positions. This info must be retained.
Another complication is, that two or more MarkerLayers can be merged to one.

All these problems are avoided by explicitly exporting the markers to a separate file (as done in this commit).


git-svn-id: http://josm.openstreetmap.de/svn/trunk@5684 0c6e7542-c601-0410-84e7-c038aed88b3b