Fix NPE and hang when adding erroneus imageries.
When user points to a getCapabilities document that is XHTML file, parser was by
default loading external entities. For XHTML, it was trying to load for ex.:
http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd - which takes about 15 seconds
to load. It referes to other external entities, that take no shorter to load,
hence was the hang.

Now factory will not try to load external resoureces, nor validate the XML, so
this should prevent such errors.

See: #12450


git-svn-id: http://josm.openstreetmap.de/svn/trunk@9664 0c6e7542-c601-0410-84e7-c038aed88b3b