use intended units for east/north coordinates (see #12186)
When east/north coordinates are not in meter, but
feet, ... convert them to correct units.
Currently they are always stored in meters or degrees.
This makes no difference to JOSM internally, but affects
services like WMS/WMTS.
Only relevant for projections with +units=... or +to_meter=...
parameter set to non-default value.


git-svn-id: http://josm.openstreetmap.de/svn/trunk@9790 0c6e7542-c601-0410-84e7-c038aed88b3b