Fix relation ordering after split-way
In some cases the split-way command failed to detect the direction of
ways part of a relation by breaking too soon from the loop that inspects
relation members. This commit fixes this.

Additionally, a warning is shown to the user if by splitting the way a
relation may have been damaged. This warning is raised when all of these
requirements are met:

* the relation concerned is a route or multipolygon,
* has more than one member,
* does not have all of its members downloaded, and
* the direction could not be determined from the available members.