Add tests for dual-carriageway rendering and sorting
This adds a few common variants of dual carriageway relations (i.e.
different ways for both directions of the relation), and verifies that
they are rendered and sorted as expected.

There are still some cases which are not sorted properly, or not in all
cases, which are marked with TODO comments in the testing code.

Testcases for relations with split ends are not added here, since those
are not yet handled properly (see #6166 and #19633) and a fix for these
issues is up next.