see #6166 - RelationNodeMap: Do not remove entry nodes for oneway ways
The deleteWayNode method is called when a way is about to be returned
and removes the way from the relevant node-to-ways lookup maps. It also
gets passed the entry node, which is removed from the way-to-nodes map
to prevent exiting the way later through the same node as it was
entered.

This is only really relevant for bidirectional ways, in the oneway
way-to-nodes maps, only exit nodes are listed anyway. In practice, this
means that this removal was almost always just a nop, but there are
corner cases (in particular involving circular ways that are not
roundabouts do have a forward or backward role) where this could leave
the way without an exit node.

I was not able to find an actual case where this caused a problem, but
by not removing the entry node for oneway ways, the rest of the code can
now safely assume that the oneway maps always contain exactly one node
(the exit node) for each way.