RelationNodeMap: Simplify end of loop checking
Previously, this would look up `exitNode` in `onewayMap.nodes`, pass the
resulting way to `processBackwardIfEndOfLoopReached`, which would then
look up the way in `onewayReverseMap.ways`, which would necessarily
return exactly the same node `exitNode` again.

This allows simplifying this code by skipping the reverse lookups and
instead just passing `exitNode` directly. Since the function would then
either returns the passed in `exitNode` (when the loop is complete) or
`null` (when the loop is not complete), it might as well just return a
boolean, which also makes the call a bit easier to read.

To reflect the slightly changed semantics and return value, the function
is also renamed to `checkIfEndOfLoopReached`.