see #6166 - RelationNodeMap: Remove redundant loop
This loops over the Node set in onewaymap.ways, but it necessarily will
always contain exactly one node, which is the exit node of the way. So
instead of looping over a set of size 1, just get the single node and
use that directly. This slightly simplifies the code, and allows
further simplification next.