RelationNodeMap: simplify popForwardOnewayPart
Now that end of loop checking no longer needs access to a possible next
way, but can just examine the exit node of the previous way, the end of
loop check can happen before finding a possible next way.

Previously, the next way was found using `findAdjacentWay` and
explicitly deleted later, because this adjacent way could be the next
forward way, but also the last backward way.

Now that the end of loop check happens before finding an adjacent way,
we can be certain that if an adjacent way is found, it is the next way
and should be returned (previously, it could also have been the last
backward way). This allows using deleteAndGetAdjacentNode, rather than
separately getting and deleting the adjacent node.

At first glance, it might look like there is a small functional change
because loop end checking is now also done even when no adjacent way is
found. However, In cases where no adjacent way is present, the code to
handle split ends would do exactly the same as the code for a loop end,
so behaviour should be unchanged in this case.