add from-to as an option for display names
Some routes (in particular node networks) have no proper name of their
own, but do have a from and/or to attribute set that gives at least some
description of the route.

Note that these routes often do have the `name` or `note` attribute set
using an explicit "from-to" value, but these attributes are not really
meant for that, so for example the Swiss hiking route tagging now
recommends settings only `from` and `to`, not `name` or `note`.

See e.g. https://wiki.openstreetmap.org/wiki/DE:Switzerland/HikingNetwork#Wanderwegenetz

As an example of such a route, see https://www.openstreetmap.org/relation/13126425

For these routes, setting just `from` or just `to` is also possible (in
particular for node network routes that start at an unnamed
node/guidepost). In this case, the missing attribute is replaced by `?`.

This commit modifies the `relation.nameOrder` preference, adding one
extra special value `from-to` which is handled specially.