Fix #17288: Notes progress bar shows no progress (patch by gaben, modified)
The patch provided by gaben was modified to only count modified notes for
the progress bar and count. Technically speaking, this is a little less
efficient, as the modified notes will have to iterate through comments
twice instead of once. This should not be noticeable to end users, as
they should not (generally speaking) be closing enough notes for this
to be an issue.

UploadNotesTask now has a test (not by gaben) to check for correctness.



git-svn-id: https://josm.openstreetmap.de/svn/trunk@18443 0c6e7542-c601-0410-84e7-c038aed88b3b