Fix #22106: Avoid allocations for point objects in QuadBuckets
Nodes do not keep their bboxes in memory (to reduce overall memory usage).
This does mean that every time someone searches a quadbucket for nodes, we
have to recreate the BBox for every node in a quadbucket leaf. This is a
significant number of allocations that are unnecessary (~80% of the memory
allocations from QuadBuckets#search for nodes is from the creation of
temporary bboxes). The remaining cost for the search is from List#grow calls.


git-svn-id: https://josm.openstreetmap.de/svn/trunk@18465 0c6e7542-c601-0410-84e7-c038aed88b3b