MainApplication: Fix CORRECTNESS NP_NULL_PARAM_DEREF
This is mostly due to a false positive (the only way
for mainJOSM to be called with null is if
`new ProgramArguments` throws an exception, but we
exit the program in that case using Lifecycle.exitJosm).

In any case, moving `mainJOSM` into the `try-catch`
block should be marginally clearer, in so far as
it is easier to tell that we don't run `mainJOSM`
if we cannot parse the program arguments.


git-svn-id: https://josm.openstreetmap.de/svn/trunk@18484 0c6e7542-c601-0410-84e7-c038aed88b3b