Preview object properties on mouse hover
Show tags and relation memberships of an object in the properties
dialog when moving the mouser pointer over it.

Can be enabled/disabled via settings (enabled by default)

An existing selection overrides the hover preview by default, but this
can be changed via a setting to always show information about hovered objects.