See #17858: start linking to Java 17 for Java updates.
The link for the Java download page now goes to azul.com, and attempts to pre-fill as
much of the download form as possible.

This also adds a method to warn users running Java 10 or earlier that their version
of Java will soon be unsupported by JOSM. This only affects users using Java 10 or
earlier if they are not running JOSM using WebStart.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18580 0c6e7542-c601-0410-84e7-c038aed88b3b