FlatLaf uses properties files which occasionally have breaking changes
In the event that a properties file is misconfigured, FlatLaf will throw
an IllegalArgumentException, which causes JOSM to exit, which is not ideal,
since recovering is not intuitive.

This will roll back the LAF to the platform default, so people can start JOSM.
Which isn't ideal, but is better than not starting indefinitely.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18598 0c6e7542-c601-0410-84e7-c038aed88b3b