Fix #20528: Copy does not work from the welcome screen
This was caused by setting the TransferHandler to `null` in order to get
drag and drop working. Instead of setting the TransferHandler to `null`, we
just set the DropTarget to `null`, which is what prevents the parent
DropTarget from being triggered.

The DropTarget could also have it's `setActive` method called with `false`,
but this will hopefully be less surprising.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18601 0c6e7542-c601-0410-84e7-c038aed88b3b