Fix #20849: WARNING: row index is bigger than sorter's row count. Most likely this is a wrong sorter usage
This is caused by the ordering of the TableModelListeners for
availableSourcesModel. While not documented, Java fires the listeners in reverse
order, so we want to remove the listener that updates the sorter and add the
column width adjustment listener, and then re-add the listener that updates the
sorter.

The alternative to adding/removing the listener for the sorter is to use an
AtomicReference for tblAvailableSources and add the adjustColumnWidth listener
before creating the tblAvailableSources.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18603 0c6e7542-c601-0410-84e7-c038aed88b3b