Fix #21423: Prevent error codes from clashing
This works by creating a unique code using the test class name. The new
format for ignore entries will look like `UNIQUECODE_CODE_MESSAGE_STATE`.
The switchover date for the new entries is set at 2024-01-01 in order to
give users sufficient time to update, such that if a user has multiple
installations of JOSM, all versions will be able to use the same ignore
list. The compatibility code is intended to be removed in 2025.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18636 0c6e7542-c601-0410-84e7-c038aed88b3b