Fix #22680: Unexpected exception when downloading GeoJSON
This occurs when we start to read the JSON, then something comes along and
interrupts the stream from the server. This then causes a JsonException
(a RuntimeException) to be thrown.

This patch catches the JsonException and then either rethrows it if the
cause was not an IOException or throws a new IllegalDataException with
the IOException as the cause and the JsonException as a suppressed exception.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18656 0c6e7542-c601-0410-84e7-c038aed88b3b