See #22770, fix failing unit tests
The fix for #22770 included a new unit test, which caused a cascading failure.
The first step of the fix included removing the destroyed dialog from the
MapFrame. Unfortunately, `MapFrame#removeToggleDialog` did not properly
remove the ToggleDialog -- it called `DialogsPanel#remove(Component)` with
the ToggleDialog as an argument. Since the DialogsPanel never adds the
ToggleDialog as a direct subcomponent, this was effectively a no-op.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18686 0c6e7542-c601-0410-84e7-c038aed88b3b