Fix #22432, see #22941: Start migrating from javax to jakarta
Parsson was split out from the JSONP repository in 2021 (see
https://github.com/jakartaee/jsonp-api/issues/285 ). It is the default provider,
and will "just work" without additional configuration.

Many plugins use `javax.json`, so the scheduled removal of the `javax.json`
dependencies is set to milestone:"24.12" (see #22941).

Changes between javax.json and jakarta.json 2.0:
* Rename of `javax.json` to `jakarta.json`
* Some additional bug fixes

This will enable us to move easily to jakarta.json 2.1 in the future.
The changes of note with 2.1 includes:
* Better handling of duplicated keys
* Additional APIs around primitive types
* API to get current event from JsonParser

We cannot currently move to jakarta.json 2.1 since it requires Java 11+.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18723 0c6e7542-c601-0410-84e7-c038aed88b3b