Fix #23001: ClassCastException in GpxDrawHelper#calculateColors
This occurs when a working GPX file with HDOP information is converted to an OSM
Data Layer and back to a GPX Data Layer. This is fixed by (a) converting casts
for `get` operations to `Number` instead of `Float`/`Integer` and (b) actually
putting the number in the attribute map instead of a string.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18753 0c6e7542-c601-0410-84e7-c038aed88b3b