Fix #23018: The Name Suggestion Index preset causes dramatic slowing in line drawing and following
The root cause is that every time the dataset fires a DataChangedEvent, we
rebuild the relation list. This causes a search in the presets, which can become
very expensive with the NSI. To fix this, we instead iterate through each
individual event in the DataChangedEvent, and perform the appropriate action.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18792 0c6e7542-c601-0410-84e7-c038aed88b3b