Fix #23140: RejectedExecutionException when MultiFetchServerObjectReader is cancelled while creating download jobs
This was caused by a race condition.
User starts a download, and cancels it as jobs are submitted to the thread pool.
The next job to be submitted will cause a RejectedExecutionException. In order
to fix this, we ensure that we shutdown the thread pool inside a synchronized
block, and add jobs to the executor inside a synchronized block.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18821 0c6e7542-c601-0410-84e7-c038aed88b3b