Charging station preset: Add fields for capacity and covered
The capacity is important, because a charging station with 4 charging
plugs may only allow two cars to charge simultaneously. The capacity
field was there already, but only allowed fixed values, which does not
make sense. There are charging stations with capacity for a dozen cars
or more.

The information, whether a station is covered, is important when
charging while it rains.