Charging station preset: Optimize vehicle classes
Remove bicycle / scooter. The whole preset is optimized for car charging
stations (after all it is grouped under "Transport > Car"). Thus, this
preset is not suited to map bicycle charging stations and including that
key does not make sense.

The "truck" key should be renamed to "hgv" as per the wiki.

I also added a short text with some guidance on what to tag.