Fix #22810: OSM OAuth 1.0a/Basic auth deprecation and removal
As of 2024-02-15, something changed in the OSM server configuration. This broke
our OAuth 1.0a implementation (see #23475). As such, we are removing OAuth 1.0a
from JOSM now instead of when the OSM server removes support in June 2024.

For third-party OpenStreetMap servers, the `Basic Authentication` method has been
kept. However, they should be made aware that it may be removed if a non-trivial
bug occurs with it. We highly recommend that the third-party servers update to
the current OpenStreetMap website implementation (if only for their own security).

Failing that, the third-party server can implement RFC8414. As of this commit,
we currently use the `authorization_endpoint` and `token_endpoint` fields.
To check and see if their third-party server implements RFC8414, they can go
to `<server host>/.well-known/oauth-authorization-server`.

Prominent third-party OpenStreetMap servers may give us a client id for their
specific server. That client id ''may'' be added to the hard-coded client id list
''at maintainer discretion''. At a minimum, the server ''must'' be publicly
available ''and'' have a significant user base.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@18991 0c6e7542-c601-0410-84e7-c038aed88b3b