Fix #23540: When a Linux distribution removes the Bing png from JMapViewer, JOSM hangs when loading a Bing imagery layer
This is due to an interaction between JOSM preferences, the ValidatorDialog, and
the cached Bing imagery layer.

The specific interaction that leads to the freeze is as follows:
1. User starts process of adding a Bing imagery layer to JOSM
2. In the ''EDT'' we call a `synchronized` method in `MainLayerManager`
3. In the ''EDT'' we block on a separate thread that is fetching the Bing
   attribution data
4. In the separate thread, we attempt to set cache information in preferences
5. Setting the preference fires a preference change event
6. One of the listeners, specifically `ValidatorDialog`, then calls another
   `synchronized` method in `MainLayerManager`

Steps 2 and 6 are the problem points. This patch fixes that by ''only''
attempting to perform the update by only listening to a specific preference,
instead of all of them.

The alternative to this patch uses `Config#addKeyPreferenceChangeListener`
instead. This was not done since `ToggleDialog` is already listening to all
preference changes.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@19016 0c6e7542-c601-0410-84e7-c038aed88b3b