Fix #21533, see #23600: Native Apple Silicon Support
When building on macOS, the macos-jpackage.sh script now takes two arguments:
1. The JOSM revision
2. An additional JDK to build against (this should be of a different architecture)

If an additional JDK is specified, it is assumed that it is a different
architecture, and the two built apps are combined using lipo for a universal binary.

`ant.yml` does have some changes to make it easier to debug packaging issues in
PRs; the PR does still need to upload the specific package and add its branch to
the triggers.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@19030 0c6e7542-c601-0410-84e7-c038aed88b3b