Fix #4142: Track fully downloaded objects (patch by stoecker, GerdP, and myself)
The serialization move from `PrimitiveData` to `AbstractPrimitive` should be
reverted prior to 24.05 (see #23677).

The serialization move was required since we want to ensure that all downstream
users of `AbstractPrimitive` were not using the `flags` field, which was done by
making the field `private` instead of `protected`. They may still be using that
field (via `updateFlags`) which would not be caught by compile-time or runtime
errors.

Additionally, a good chunk of common functionality was moved up from
`OsmPrimitive`, even though much of it wasn't useful for `PrimitiveData`.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@19078 0c6e7542-c601-0410-84e7-c038aed88b3b