See #23671: Significantly improve performance of Utils.isBlank and Utils.isStripEmpty along with reducing intern cost when loading files
From `Utils`, `isBlank` and `isStripEmpty` are functionally similar; both take a
string and check for the same whitespace characters. Specifically, the code flow
for `isBlank` calls `strip`, which uses `DEFAULT_STRIP` as additional characters
to remove. `strip` eventually calls `isStrippedChar` to determine whether or not
to keep a character. `isStrippedChar` checks `DEFAULT_STRIP` always, so the
returns from `Utils.isBlank` and `Utils.isStripEmpty` should always be the same.

Since the two functions are the same, we change `isBlank` to point at
`isStripEmpty`, and `isStripEmpty` now calls `String.isBlank` prior to checking
every character. `String.isBlank` is a built-in method, and is thus more likely
to be well-optimized by the JVM. In the event that it returns false, we want to
avoid any additional work, so we use a traditional `for` loop instead of a stream;
we will usually know within a character or two whether or not the string is empty.
With a stream, we first have to construct the stream, then iterate through it.

In `AbstractReader`, we create a map of strings to interned strings to avoid
calling `String.intern` too often -- despite being something we want to do fairly
frequently. Part of the (current) problem with String.intern is that it uses a
`ConcurrentHashTable`, so it "costs" more to use. The `HashMap` (even with the
cost of the lambdas) is ~25% of the CPU cost of interning, given sufficiently
duplicated data.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@19079 0c6e7542-c601-0410-84e7-c038aed88b3b