Fix #8269: Add initial maven pom files
This additionally deletes ''most'' IDE specific configuration, in favor of Maven
configuration.

We did have to update scripts/BuildProjectionDefinitions.java so that it would
create an empty EPSG file for the next run in the pom.xml file.

For core, we needed two additional pom.xml files:
* A parent pom with most dependency information (for inheritance)
 * This will make it easier to keep core and the apache-commons plugin in sync
* A unit test pom (for plugin tests)
 * Some IDEs (netbeans specifically) don't like this.

What this does ''not'' do:
* Convert from ivy to maven for dependency management in core. This should happen
  eventually with `maven-resolver-ant-tasks` (probably with the next dependency
  update).
* Deprecate the current `ant` scripts.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@19097 0c6e7542-c601-0410-84e7-c038aed88b3b