Fix a performance issue when working with large datasets
The primary issue is that the collection passed to the resetTiles method is a
filtered collection with no defined size. In order to calculate the size, it must
iterate through the collection which becomes expensive with large datasets.

The fix for this is to remove the dependency on the size of the dataset, and just
look at the number of primitives passed in.

In addition, if the current renderer is ''not'' StyledTiledMapRenderer, clear the cache and return.

git-svn-id: https://josm.openstreetmap.de/svn/trunk@19271 0c6e7542-c601-0410-84e7-c038aed88b3b