See #24046: Reduce cost of Geometry#filterInsidePolygon when a primitive is a relation
This was done by checking that the bounds of the polygon contain the bounds of
the multipolygon ring before creating an area from the multipolygon ring.

If the polygon does ''not'' contain the bounds of the ring, then at least part
of the ring is outside the polygon, and thus is not inside the polygon.

Measurements using validators at (57.5183581,-75.0512982),(57.2181217,-73.9434821):
* CPU: -87%
* Memory: -62%

Please note that the test area is still very expensive in
Geometry#filterInsideMultipolygon (note the `Multipolygon`).


git-svn-id: https://josm.openstreetmap.de/svn/trunk@19272 0c6e7542-c601-0410-84e7-c038aed88b3b