Fix performance regression when hovering over primitives
SVN 19176 introduced a performance regression in the OSM data layer where
hovering over a primitive triggers a layer invalidation and tile cache reset,
all presumably for the tiled renderer. When working with even modest datasets,
this triggers superfluous repaints of the MapView. The only active renderer
check is in `resetTiles` (SVN 19271)—and this fails to guard against any other
execution specific to the tiled renderer.